+++ PusherClient/IChannelDataDecrypter.cs
@@ -1,0 +1,7 @@
+﻿namespace PusherClient
+{
+    internal interface IChannelDataDecrypter
+    {
+        string DecryptData(byte[] decryptionKey, EncryptedChannelData encryptedData);
+    }
+}